+++ core/src/it/dukemania/midi/PercussionNote.java
@@ -6,7 +6,7 @@ public class PercussionNote extends AbstractNote {
     private static final int OFFSET = 35;
     private final Percussion instrument;
 
-    public PercussionNote(final Optional<Double> duration, final int startTime, final int identifier) {
+    public PercussionNote(final Optional<Long> duration, final long startTime, final int identifier) {
         super(duration, startTime, identifier);
         this.instrument = Percussion.values()[identifier - OFFSET];
     }
